Consolidate trustees function handler
Move the software-trustees Azure Function handler into
bankruptcy-software.function.ts instead of having a separate
module, matching the pattern used for the name endpoint.

Jira ticket: CAMS-669
